Add constructor to Benchmark class
The array size and increase currently used to run the algorithm against 
is hard-coded within the test function. This currently prevents anyway 
to run against different array length step increases or up to a 
different maximum length.

Add parameters to constructor of class so that the step increase and 
maximum array length can be passed as arguements when Benchmark instance 
is created.

Defaults should be specified at 50,000 for step increases and 1,000,000 
for the maximum.